backend/requirements.txt
pandas
requests
flask
flask_cors
python-dotenv